+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 N/A
 
 ### Fixes
-N/A
+- When on_schema_change is set, pass common columns as dest_columns in incremental merge macros ([#4144](https://github.com/dbt-labs/dbt-core/issues/4144))
 
 ### Under the hood
 - Add optional profile parameters for atypical local connection setups ([#21](https://github.com/dbt-labs/dbt-snowflake/issues/21), [#36](https://github.com/dbt-labs/dbt-snowflake/pull/36))
@@ -19,6 +19,7 @@ N/A
 
 ### Contributors
 - [@NiallRees](https://github.com/NiallRees) ([#32](https://github.com/dbt-labs/dbt-snowflake/pull/32))
+- [@Kayrnt](https://github.com/Kayrnt) ([38](https://github.com/dbt-labs/dbt-snowflake/pull/38))
 
 ## dbt-snowflake v1.0.0b1 (October 11, 2021)